Move LifecyclePolicyTestsUtils (again)
It seems more at home with all its callers in
x-pack/plugin/ilm/src/test
