+++ x-pack/plugin/ilm/src/test/java/org/elasticsearch/xpack/ilm/LockableLifecycleType.java
@@ -3,9 +3,12 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-package org.elasticsearch.xpack.core.ilm;
+package org.elasticsearch.xpack.ilm;
 
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.xpack.core.ilm.LifecycleAction;
+import org.elasticsearch.xpack.core.ilm.LifecycleType;
+import org.elasticsearch.xpack.core.ilm.Phase;
 
 import java.util.ArrayList;
 import java.util.Collection;
